requirements.txt
# Core ML/AI libraries
torch>=2.0.0
transformers>=4.30.0
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# UI Framework
gradio>=4.0.0

# Vector Database and Embeddings
chromadb>=0.4.0
numpy>=1.21.0

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.0
unstructured>=0.10.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Optional: OpenAI compatibility
openai>=1.0.0

# Development tools (optional)
black>=23.0.0
pytest>=7.0.0
flake8>=6.0.0

# Additional utilities
tqdm>=4.65.0
rich>=13.0.0